Refactor dataflow framework. Introduce backward analysis.
