+++ changelog.txt
@@ -7,13 +7,21 @@ The Nullness Checker now treats `System.getProperty()` soundly.  Use
 `-Alint=permitClearProperty` to disable special treatment of
 `System.getProperty()` and to permit undefining built-in system properties.
 
+The Dataflow Framework now also supports backward analysis.
+
 Implementation details:
-Changed the types of some fields and methods from array to List:
-QualifierDefaults.validLocationsForUncheckedCodeDefaults()
-QualifierDefaults.STANDARD_CLIMB_DEFAULTS_TOP
-QualifierDefaults.STANDARD_CLIMB_DEFAULTS_BOTTOM
-QualifierDefaults.STANDARD_UNCHECKED_DEFAULTS_TOP
-QualifierDefaults.STANDARD_UNCHECKED_DEFAULTS_BOTTOM
+ * Changed the types of some fields and methods from array to List:
+   QualifierDefaults.validLocationsForUncheckedCodeDefaults()
+   QualifierDefaults.STANDARD_CLIMB_DEFAULTS_TOP
+   QualifierDefaults.STANDARD_CLIMB_DEFAULTS_BOTTOM
+   QualifierDefaults.STANDARD_UNCHECKED_DEFAULTS_TOP
+   QualifierDefaults.STANDARD_UNCHECKED_DEFAULTS_BOTTOM
+ * Dataflow Framework: Analysis is now an interface, added AbstractAnalysis,
+   ForwardAnalysis, ForwardTransferFunction, ForwardAnalysisImpl,
+   BackwardAnalysis, BackwardTransferFunction, BackwardAnalysisImpl.
+   Existing code is easy to adapt:
+     `extends Analysis<V, S, T>` => `extends ForwardAnalysisImpl<V, S, T>`
+     `implements TransferFunction<V, S>` => `implements ForwardTransferFunction<V, S>`
 
 ---------------------------------------------------------------------------
