+++ dataflow/src/main/java/org/checkerframework/dataflow/analysis/RegularTransferResult.java
@@ -86,7 +86,7 @@ public class RegularTransferResult<V extends AbstractValue<V>, S extends Store<S
      *
      * <p><em>Aliasing</em>: {@code resultStore} and any store in {@code exceptionalStores} are not
      * allowed to be used anywhere outside of this class (including use through aliases). Complete
-     * control over the objects is transfered to this class.
+     * control over the objects is transferred to this class.
      *
      * @param value the abstract value produced by the transfer function
      * @param resultStore {@link #store}
