+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/AbstractCFGVisualizer.java
@@ -15,6 +15,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
 import org.checkerframework.dataflow.analysis.AbstractValue;
 import org.checkerframework.dataflow.analysis.Analysis;
+import org.checkerframework.dataflow.analysis.Analysis.Direction;
 import org.checkerframework.dataflow.analysis.Store;
 import org.checkerframework.dataflow.analysis.TransferFunction;
 import org.checkerframework.dataflow.analysis.TransferInput;
@@ -25,6 +26,7 @@ import org.checkerframework.dataflow.cfg.block.RegularBlock;
 import org.checkerframework.dataflow.cfg.block.SingleSuccessorBlock;
 import org.checkerframework.dataflow.cfg.block.SpecialBlock;
 import org.checkerframework.dataflow.cfg.node.Node;
+import org.checkerframework.javacutil.BugInCF;
 
 /**
  * This abstract class makes implementing a {@link CFGVisualizer} easier. Some of the methods in
@@ -196,26 +198,12 @@ public abstract class AbstractCFGVisualizer<
 
         // Visualize transfer input if necessary.
         if (analysis != null) {
-            // The transfer input before this block is added before the block content.
-            sbBlock.insert(0, visualizeBlockTransferInput(bb, analysis));
+            sbBlock.insert(0, visualizeBlockTransferInputBefore(bb, analysis));
             if (verbose) {
-                Node lastNode = getLastNode(bb);
-                if (lastNode != null) {
-                    @SuppressWarnings("nullness:contracts.precondition.not.satisfied")
-                    S store = analysis.getResult().getStoreAfter(lastNode);
-                    StringBuilder sbStore = new StringBuilder();
-                    sbStore.append(escapeString).append("~~~~~~~~~").append(escapeString);
-                    sbStore.append("After: ");
-                    if (store != null) {
-                        sbStore.append(visualizeStore(store));
-                    } else {
-                        sbStore.append("null store");
-                    }
-                    sbBlock.append(sbStore);
-                }
+                sbBlock.append(visualizeBlockTransferInputAfter(bb, analysis));
             }
         }
-        if (!centered) {
+        if (!centered || verbose) {
             sbBlock.append(escapeString);
         }
         return sbBlock.toString();
@@ -253,16 +241,15 @@ public abstract class AbstractCFGVisualizer<
             case EXCEPTION_BLOCK:
                 return Collections.singletonList(((ExceptionBlock) bb).getNode());
             case CONDITIONAL_BLOCK:
-                return Collections.emptyList();
             case SPECIAL_BLOCK:
                 return Collections.emptyList();
             default:
-                throw new Error("Unrecognized basic block type: " + bb.getType());
+                throw new BugInCF("Unrecognized basic block type: " + bb.getType());
         }
     }
 
     /**
-     * Visualize the transfer input of a block.
+     * Visualize the transfer input before a block.
      *
      * @param bb the block
      * @param analysis the current analysis
@@ -271,30 +258,98 @@ public abstract class AbstractCFGVisualizer<
      *     StringCFGVisualizer} to simply add a new line
      * @return the String representation of the transfer input of the block
      */
-    protected String visualizeBlockTransferInputHelper(
+    protected String visualizeBlockTransferInputBeforeHelper(
             Block bb, Analysis<V, S, T> analysis, String escapeString) {
-        assert analysis != null
-                : "analysis should be non-null when visualizing the transfer input of a block.";
 
-        TransferInput<V, S> input = analysis.getInput(bb);
-        assert input != null : "@AssumeAssertion(nullness): well-behaved analysis";
+        if (analysis == null) {
+            throw new BugInCF(
+                    "analysis must be non-null when visualizing the transfer input of a block.");
+        }
+
+        S regularStore;
+        S thenStore = null;
+        S elseStore = null;
+        boolean isTwoStores = false;
 
         StringBuilder sbStore = new StringBuilder();
-
-        // Split input representation to two lines.
         sbStore.append("Before: ");
-        if (!input.containsTwoStores()) {
-            S regularStore = input.getRegularStore();
+
+        Direction analysisDirection = analysis.getDirection();
+
+        if (analysisDirection == Direction.FORWARD) {
+            TransferInput<V, S> input = analysis.getInput(bb);
+            assert input != null : "@AssumeAssertion(nullness): invariant";
+            isTwoStores = input.containsTwoStores();
+            regularStore = input.getRegularStore();
+            thenStore = input.getThenStore();
+            elseStore = input.getElseStore();
+        } else {
+            regularStore = analysis.getResult().getStoreBefore(bb);
+        }
+
+        if (!isTwoStores) {
             sbStore.append(visualizeStore(regularStore));
         } else {
-            S thenStore = input.getThenStore();
+            assert thenStore != null : "@AssumeAssertion(nullness): invariant";
+            assert elseStore != null : "@AssumeAssertion(nullness): invariant";
             sbStore.append("then=");
             sbStore.append(visualizeStore(thenStore));
-            S elseStore = input.getElseStore();
             sbStore.append(", else=");
             sbStore.append(visualizeStore(elseStore));
         }
         sbStore.append(escapeString).append("~~~~~~~~~").append(escapeString);
+        return sbStore.toString();
+    }
+
+    /**
+     * Visualize the transfer input after the block.
+     *
+     * @param bb the block
+     * @param analysis the current analysis
+     * @param escapeString the escape String for the special need of visualization, e.g., "\\l" for
+     *     {@link DOTCFGVisualizer} to keep line left-justification, "\n" for {@link
+     *     StringCFGVisualizer} to simply add a new line
+     * @return the String representation of the transfer input of the block
+     */
+    protected String visualizeBlockTransferInputAfterHelper(
+            Block bb, Analysis<V, S, T> analysis, String escapeString) {
+        if (analysis == null) {
+            throw new BugInCF(
+                    "analysis should be non-null when visualizing the transfer input of a block.");
+        }
+
+        S regularStore;
+        S thenStore = null;
+        S elseStore = null;
+        boolean isTwoStores = false;
+
+        StringBuilder sbStore = new StringBuilder();
+        sbStore.append("After: ");
+
+        Direction analysisDirection = analysis.getDirection();
+
+        if (analysisDirection == Direction.FORWARD) {
+            regularStore = analysis.getResult().getStoreAfter(bb);
+        } else {
+            TransferInput<V, S> input = analysis.getInput(bb);
+            assert input != null : "@AssumeAssertion(nullness): invariant";
+            isTwoStores = input.containsTwoStores();
+            regularStore = input.getRegularStore();
+            thenStore = input.getThenStore();
+            elseStore = input.getElseStore();
+        }
+
+        if (!isTwoStores) {
+            sbStore.append(visualizeStore(regularStore));
+        } else {
+            assert thenStore != null : "@AssumeAssertion(nullness): invariant";
+            assert elseStore != null : "@AssumeAssertion(nullness): invariant";
+            sbStore.append("then=");
+            sbStore.append(visualizeStore(thenStore));
+            sbStore.append(", else=");
+            sbStore.append(visualizeStore(elseStore));
+        }
+        sbStore.insert(0, escapeString + "~~~~~~~~~" + escapeString);
         return sbStore.toString();
     }
 
@@ -314,28 +369,7 @@ public abstract class AbstractCFGVisualizer<
             case EXCEPTIONAL_EXIT:
                 return "<exceptional-exit>" + separator;
             default:
-                throw new Error("Unrecognized special block type: " + sbb.getType());
-        }
-    }
-
-    /**
-     * Returns the last node of a block, or null if none.
-     *
-     * @param bb the block
-     * @return the last node of this block or {@code null}
-     */
-    protected @Nullable Node getLastNode(Block bb) {
-        switch (bb.getType()) {
-            case REGULAR_BLOCK:
-                List<Node> blockContents = ((RegularBlock) bb).getContents();
-                return blockContents.get(blockContents.size() - 1);
-            case CONDITIONAL_BLOCK:
-            case SPECIAL_BLOCK:
-                return null;
-            case EXCEPTION_BLOCK:
-                return ((ExceptionBlock) bb).getNode();
-            default:
-                throw new Error("Unrecognized block type: " + bb.getType());
+                throw new BugInCF("Unrecognized special block type: " + sbb.getType());
         }
     }
