+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/playground/ConstantPropagationPlayground.java
@@ -1,6 +1,7 @@
 package org.checkerframework.dataflow.cfg.playground;
 
-import org.checkerframework.dataflow.analysis.Analysis;
+import org.checkerframework.dataflow.analysis.ForwardAnalysis;
+import org.checkerframework.dataflow.analysis.ForwardAnalysisImpl;
 import org.checkerframework.dataflow.cfg.CFGVisualizeLauncher;
 import org.checkerframework.dataflow.constantpropagation.Constant;
 import org.checkerframework.dataflow.constantpropagation.ConstantPropagationStore;
@@ -19,10 +20,10 @@ public class ConstantPropagationPlayground {
 
         // run the analysis and create a PDF file
         ConstantPropagationTransfer transfer = new ConstantPropagationTransfer();
-        Analysis<Constant, ConstantPropagationStore, ConstantPropagationTransfer> analysis =
-                new Analysis<>(transfer);
+        ForwardAnalysis<Constant, ConstantPropagationStore, ConstantPropagationTransfer>
+                forwardAnalysis = new ForwardAnalysisImpl<>(transfer);
         CFGVisualizeLauncher cfgVisualizeLauncher = new CFGVisualizeLauncher();
         cfgVisualizeLauncher.generateDOTofCFG(
-                inputFile, outputDir, method, clazz, true, false, analysis);
+                inputFile, outputDir, method, clazz, true, false, forwardAnalysis);
     }
 }
